requirements.txt
# API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# AI / ML
open-clip-torch==2.24.0
torch>=2.9.0
torchvision>=0.20.0
Pillow>=10.3.0

# Vector Database
qdrant-client>=1.9.1

# Task Queue
redis>=5.0.4
rq>=1.16.2

# Utilities
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.0
httpx>=0.27.0
numpy>=1.26.4
